+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## unreleased
 
+- feat: renamed -s/--source to -f/--format and -S/--style to -s/--style
+- cli: fix action warnings in GitHub workflow
+
 ## v0.2.0 (2026-06-13)
 
 - feat: added --prettify flag for re-formatting existing tables